requirements.txt
exif==1.3.5
Pillow==9.3.0
Send2Trash==1.8.0
